+++ src/commands/index.ts
@@ -1,0 +1,4 @@
+import { Command } from '@/lib/wa'
+import { pingCommand } from './ping'
+
+export const commands: Command[] = [pingCommand]
