+++ src/lib/server/index.ts
@@ -1,8 +1,8 @@
+import { routes } from '@/routes'
 import cors from 'cors'
-import express from 'express'
+import express, { RequestHandler } from 'express'
 import http from 'http'
 import { Server } from 'socket.io'
-import { router } from './routes'
 
 export const app = express()
 export const server = http.createServer(app)
@@ -10,4 +10,28 @@ export const io = new Server(server, { cors: { origin: '*' } })
 
 app.use(cors())
 app.use(express.json())
-app.use(router)
+
+export type Route = {
+  path: string
+  method: string
+  handler: RequestHandler
+}
+
+routes.forEach((route) => {
+  switch (route.method) {
+    case 'GET':
+      app.get(route.path, route.handler)
+      break
+    case 'POST':
+      app.post(route.path, route.handler)
+      break
+    case 'PUT':
+      app.put(route.path, route.handler)
+      break
+    case 'DELETE':
+      app.delete(route.path, route.handler)
+      break
+    default:
+      break
+  }
+})
