+++ src/routes/index.ts
@@ -1,0 +1,5 @@
+import { Route } from '@/lib/server'
+import { rootRoute } from './root'
+import { sendRoute } from './send'
+
+export const routes: Route[] = [rootRoute, sendRoute]
